[#123] feature: Delete start button from guest player canvas and rename ViewController name
