+++ Assets/Scripts/UIController/Title/TitleViewModel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c6a7750787d434cf0ae8fff77e3cfc7f
+guid: 9a9a903f036514984925d01f23b43180
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
